backend/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-multipart>=0.0.6
langchain-openai>=0.0.2
langchain>=0.0.350
pinecone-client>=2.2.4
supabase>=2.0.2
manim>=0.19.0
numpy>=1.24.3
Pillow>=9.5.0
latex>=0.7.0
pylatex>=1.4.1
pydantic>=2.11.4
pydantic-settings>=2.9.1
python-dotenv>=1.1.0
beautifulsoup4>=4.13.4
lxml>=5.4.0
boto3>=1.34.0
langchain-pinecone
langchain-community
langchain-cohere